+++ scripts/iriunwebcam.sh
@@ -6,13 +6,15 @@ bash <(wget -qO- https://raw.githubusercontent.com/rauldipeas/apt-repository/mai
 pkcon install deb-get
 
 # Instalação do GCC 12
-pkcon install gcc-12
-sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-12 12
-sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-11 11
+if [ -f /usr/bin/mainline ]; then
+    pkcon install gcc-12
+    sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-12 12
+    sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-11 11
 
-# Instalação do v4l2loopback-dkms
-wget -qc --show-progress http://mirrors.kernel.org/ubuntu/pool/universe/v/v4l2loopback/$(wget -qO- http://mirrors.kernel.org/ubuntu/pool/universe/v/v4l2loopback/|grep v4l2loopback-dkms|grep .deb|tail -n4|head -n1|cut -d '"' -f2)
-pkcon install-local ./v4l2loopback-dkms*.deb
+    # Instalação do v4l2loopback-dkms
+    wget -qc --show-progress http://mirrors.kernel.org/ubuntu/pool/universe/v/v4l2loopback/$(wget -qO- http://mirrors.kernel.org/ubuntu/pool/universe/v/v4l2loopback/|grep v4l2loopback-dkms|grep .deb|tail -n4|head -n1|cut -d '"' -f2)
+    pkcon install-local ./v4l2loopback-dkms*.deb
+fi
 
 # Instalação do Iriun Webcam
 deb-get install iriunwebcam
