+++ scripts/webp-pixbuf-loader.sh
@@ -2,6 +2,8 @@
 set -e
 
 # Miniaturas de imagem webp no XFCE
-sudo add-apt-repository ppa:helkaluin/webp-pixbuf-loader
-pkcon install webp-pixbuf-loader
-rm -rfv "$HOME"/.cache/thumbnails
+if [ -f /usr/share/xsessions/xubuntu.desktop ]; then
+    sudo add-apt-repository ppa:helkaluin/webp-pixbuf-loader
+    pkcon install webp-pixbuf-loader
+    rm -rfv "$HOME"/.cache/thumbnails
+fi
